#111 Separate each files from `gatsbyNode/index.ts`
