+++ src/gatsbyNode/index.ts
@@ -1,79 +1,7 @@
-import path from 'path';
-import { createFilePath } from 'gatsby-source-filesystem';
-import { GatsbyNode } from 'gatsby';
+import createPages from './createPages';
+import onCreateNode from './onCreateNode';
 
-import parseArticlePath from './parseArticlePath';
-import { MarkdownRemarkConnection } from '../../types/graphqlTypes';
-
-const query = `
-    query PagesQuery {
-        allMarkdownRemark(filter: { frontmatter: { publish: { ne: false } } }) {
-            edges {
-                node {
-                    fields {
-                        slug
-                    }
-                }
-            }
-        }
-    }
-`;
-
-type Result = {
-    allMarkdownRemark: MarkdownRemarkConnection,
-};
-
-/**
- * onCreateNode
- */
-export const onCreateNode: GatsbyNode['onCreateNode'] = (args) => {
-    const { node, getNode, actions } = args;
-    const { createNodeField } = actions;
-
-    if (node.internal.type === 'MarkdownRemark') {
-        const relativeFilePath = createFilePath({
-            node,
-            getNode,
-            basePath: 'articles',
-        });
-
-        createNodeField({
-            node,
-            name: 'slug',
-            value: `/articles${relativeFilePath}`,
-        });
-
-        const [
-            date,
-            index,
-            name,
-        ] = parseArticlePath(relativeFilePath);
-
-        createNodeField({ node, name: 'date', value: date });
-        createNodeField({ node, name: 'index', value: index });
-        createNodeField({ node, name: 'name', value: name });
-    }
-};
-
-export const createPages: GatsbyNode['createPages'] = async ({ graphql, actions }) => {
-    const { createPage } = actions;
-
-    const result = await graphql<Result>(query);
-    if (result.errors) throw result.errors;
-
-    const nodes = result?.data?.allMarkdownRemark.edges;
-    if (nodes == null) return;
-
-    nodes.forEach(({ node }) => {
-        const nodePath = node?.fields?.slug;
-        if (nodePath == null) throw new Error('node.field.slug should not be nullish');
-
-        createPage({
-            path: nodePath,
-            component: path.resolve('./src/templates/ArticlePage.tsx'),
-            context: {
-                slug: nodePath,
-            },
-        });
-    });
+export {
+    createPages,
+    onCreateNode,
 };
